refactor: clean up technical debt - remove deprecated code and dependencies
DEBT-001 RESOLVED: Remove empty placeholder directories
- Deleted src/database_services/ and src/retrieval_factory/
- Microsoft Agent Framework provides pluggable memory modules

DEBT-002 RESOLVED: Remove deprecated LangGraph orchestrator
- Deleted src/orchestrators/langgraph_orchestrator.py (deprecated)
- Deleted src/agents/graph/ directory (nodes.py, workflow.py, state.py)
- Deleted src/agents/retrieval_agent.py (dead code, never wired - #134)
- Deleted src/tools/web_search.py (dead code)
- Removed LangGraph/LangChain/duckduckgo-search dependencies

Model preservation:
- Moved Hypothesis and Conflict models from graph/state.py to utils/models.py
- These shared domain models are used by ResearchMemory
- Added new test file tests/unit/utils/test_cognitive_models.py

Dependency cleanup:
- Removed 21 packages: langgraph, langchain, langchain-core,
  langchain-huggingface, langgraph-checkpoint-sqlite, duckduckgo-search,
  and their transitive dependencies

Documentation updates:
- Updated debt-registry.md (2 items resolved, dashboard updated)
- Updated overview.md (removed LangGraph from architecture diagram)

All 309 unit tests pass.
